Fixes to specific cases of find references
