+++ Source/DafnyCore/AST/Members/Function.cs
@@ -133,7 +133,9 @@ public class Function : MemberDecl, TypeParameter.ParentType, ICallable, ICanFor
     Concat<Node>(Req).
     Concat(Ens.Select(e => e.E)).
     Concat(Decreases.Expressions).
-    Concat(Formals).Concat(Result != null ? new List<Node>() { Result } : new List<Node>()).
+    Concat(Formals).
+    Concat(Result != null ? new List<Node>() { Result } : new List<Node>()).
+    Concat(ResultType != null ? new List<Node>() { ResultType } : new List<Node>()).
     Concat(Body == null ? Enumerable.Empty<Node>() : new[] { Body });
 
   public override IEnumerable<INode> PreResolveChildren =>
