+++ Source/DafnyCore/AST/Modules/AliasModuleDecl.cs
@@ -9,7 +9,7 @@ namespace Microsoft.Dafny;
 /// <summary>
 /// Represents "module name = path;", where name is an identifier and path is a possibly qualified name.
 /// </summary>
-public class AliasModuleDecl : ModuleDecl, IHasUsages, ICanFormat {
+public class AliasModuleDecl : ModuleDecl, ICanFormat {
   public readonly ModuleQualifiedId TargetQId; // generated by the parser, this is looked up
   public readonly List<IToken> Exports; // list of exports sets
   [FilledInDuringResolution] public bool ShadowsLiteralModule;  // initialized early during Resolution (and used not long after that); true for "import opened A = A" where "A" is a literal module in the same scope
@@ -41,9 +41,6 @@ public class AliasModuleDecl : ModuleDecl, IHasUsages, ICanFormat {
   }
 
   public override ModuleDefinition Dereference() { return Signature.ModuleDef; }
-  public IEnumerable<IDeclarationOrUsage> GetResolvedDeclarations() {
-    return new[] { Dereference() };
-  }
 
   public bool SetIndent(int indentBefore, TokenNewIndentCollector formatter) {
     if (OwnedTokens.FirstOrDefault() is { } theToken) {
@@ -52,4 +49,6 @@ public class AliasModuleDecl : ModuleDecl, IHasUsages, ICanFormat {
 
     return true;
   }
+
+  public override IEnumerable<INode> Children => base.Children.Concat(new INode[] { TargetQId });
 }
